requirements.txt
# Binance API
binance-futures-connector>=3.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# HTTP/WebSocket
requests>=2.31.0
websocket-client>=1.6.0

# Telegram notifications (optional)
python-telegram-bot>=20.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Type hints
typing-extensions>=4.5.0

# Utilities
python-dotenv>=1.0.0